dapr-kafka/app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Hammad-Rehman <124163466+developer-hammad-rehman@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.112.2"
uvicorn = "^0.30.6"
dapr-ext-fastapi = "^1.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
